+++ packages/backend/src/@types/knex-tables.d.ts
@@ -6,10 +6,12 @@ import {
     OrganizationTableName,
     OrganizationTable,
 } from '../database/entities/organizations';
+import { UserTable, UserTableName } from '../database/entities/users';
 
 declare module 'knex/types/tables' {
     interface Tables {
         [InviteLinkTableName]: InviteLinkTable;
         [OrganizationTableName]: OrganizationTable;
+        [UserTableName]: UserTable;
     }
 }